refactor : 기능 분리를 위한 클래스명 리팩토링
기존의 모든 비지니스 로직을 가지고 있던 RacingGame 클래스를 전체 게임 과정을 컨트롤하는 GameController로 수정
모든 자동차 객체를 가지고 전진 또는 정지 작업을 관리하는 CarController 클래스 추가
