+++ src/main/java/racingcar/Application.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class Application {
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
-        RacingGame game = new RacingGame(scanner);
-        game.playingGame();
+        GameController game = new GameController(scanner);
+        game.controlGame();
     }
 }
