+++ src/main/java/racingcar/CarController.java
@@ -1,0 +1,4 @@
+package racingcar;
+
+public class CarController {
+}
